ver 0-8-0 (work on saving new CheckList AGAIN)

+++ app/src/main/java/com/example/letscheck/data/classes/input/NewEntity.kt
@@ -42,7 +42,7 @@ data class NewEntity(
         checkBoxTitles[index] = mutableList.toList()
     }
     
-    fun renameCheckBoxTitle(index: Int, checkBoxId: Int, str: String) {
+    fun renameCheckBoxTitle(index: Int, checkBoxId: Long, str: String) {
         checkBoxTitles[index].find { it.id == checkBoxId }!!.text = str
     }
